+++ src/postorius/tests/test_forms.py
@@ -16,7 +16,7 @@
 # Postorius.  If not, see <http://www.gnu.org/licenses/>.
 from django.utils import unittest
 
-from postorius.forms import UserPreferences, DomainNew
+from postorius.forms import UserPreferences, DomainNew, ListNew
 
 
 class UserPreferencesTest(unittest.TestCase):
@@ -49,3 +49,26 @@ class DomainNewTest(unittest.TestCase):
             'contact_address': 'contact@mailman.most-desirable.org',
         })
         self.assertFalse(form.is_valid())
+
+class ListNewTest(unittest.TestCase):
+
+    def test_form_fields_list(self):
+        form = ListNew({
+            'listname': 'xyz',
+            'mail_host': 'mailman.most-desirable.org',
+            'list_owner': 'contact@mailman.most-desirable.org',
+            'advertise': 'abcd',
+            'description': 'The Most Desirable organization',
+        })
+        self.assertTrue(form.is_valid)
+
+    def test_form_fields_list_invalid(self):
+        form = ListNew({
+             'listname': 'xy#z',
+             'mail_host': 'mailman.most-desirable.org',
+             'list_owner': 'mailman.most-desirable.org',
+             'advertise': 'abcd',
+             'description': 'The Most Desirable organization',
+        })
+        self.assertFalse(form.is_valid())
+