+++ src/postorius/doc/news.rst
@@ -22,7 +22,7 @@ along with Postorius. If not, see <http://www.gnu.org/licenses/>.
 
 1.0 beta 2
 ==========
-(2015-xx-xx)
+(2015-04-15)
 
 * French translation. Provided by Guillaume Libersat
 * Addedd an improved test harness using WebTest. Contributed by Aurélien Bompard.
@@ -37,6 +37,8 @@ along with Postorius. If not, see <http://www.gnu.org/licenses/>.
 * Rework of internal testing
 * Mozilla Persona integration: switch from django-social-auto to django-browserid: Contributed by Abhilash Raj.
 * Fix manage.py mmclient command for non-IPython shells. Contributed by Ankush Sharma (LP: 1428169).
+* Added archiver options: Site-wide enabled archivers can not be enabled 
+on a per-list basis through the web UI.
 
 
 1.0 beta 1 -- "Year of the Parrot"